+++ .gitignore
@@ -3,3 +3,5 @@ bin/
 *.classpath
 *.project
 target/
+.idea
+ParticleEffect.iml
